Added new shooter subsystem and shoot only command
New ShooterBase subsystem for configuring and adding
new commands.
New ShootOnly command: This takes a value from the manipulator on the
joystick and maps it to a number 0-5000 (according to Matthew) the
lowest and highest setting for the motor.
Added new JoyStick object in OI.
Initialized SHOOTER_JOYSTICK_ID.
Initialized SHOOTER_WHEEL_MOTOR_ID to 0 because we do not know what
Talon will be controlling it.
Instantiated new ShooterBase in Robot.